First sweeping commit
Remove all items that don't directly produce the library.
This was done to simplify the CMakeLists.txt .
Also helps keeps things easy to follow.
Can look upstream if those items are needed.

Rework the CMakeLists.txt to create a static lib in a layout that will integrate with devkitpro.
The design for which was taken from Grrlib. Which looks less advanced that Box2d's, but should work fine